Refine the catalog struct layout
