+++ query/src/catalogs/table_memory_meta.rs
