+++ query/src/datasources/database/fuse/mod.rs
@@ -13,4 +13,6 @@
 //  limitations under the License.
 //
 
-pub mod database;
+mod database;
+
+pub use database::FuseDatabase;
