+++ query/src/tests/context.rs
@@ -21,8 +21,8 @@ use common_meta_types::NodeInfo;
 use crate::clusters::Cluster;
 use crate::configs::Config;
 use crate::datasources::context::DataSourceContext;
-use crate::datasources::database_engine_registry::DatabaseEngineRegistry;
-use crate::datasources::table_engine_registry::TableEngineRegistry;
+use crate::datasources::DatabaseEngineRegistry;
+use crate::datasources::TableEngineRegistry;
 use crate::sessions::DatabendQueryContext;
 use crate::sessions::DatabendQueryContextRef;
 use crate::sessions::DatabendQueryContextShared;
